optimize:optimize logging collection config
